requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.5
tensorflow==2.3.1